+++ lib/jsapi/controller/error_result.rb
@@ -2,18 +2,25 @@
 
 module Jsapi
   module Controller
-    # Used by +api_operation!+ to produce an error response.
+    # Used by Methods#api_operation! to produce an error response.
     class ErrorResult
+      delegate_missing_to :@exception
 
       # The HTTP status code of the error response to be produced.
       attr_reader :status
 
-      delegate :message, :to_s, to: :@exception
-
       # Creates a new instance to produce an error response with the given HTTP status code.
-      def initialize(exception, status:)
+      def initialize(exception, status: nil)
         @exception = exception
         @status = status
+      end
+
+      # Returns the string representation of the exception encountered to render this string
+      # when the response type is +string+, for example:
+      #
+      #   response 500, type: 'string'
+      def to_s
+        @exception.to_s
       end
     end
   end