+++ lib/jsapi/controller/parameters_invalid.rb
@@ -2,8 +2,10 @@
 
 module Jsapi
   module Controller
-    # Raised by Methods#api_operation! if the request parameters are invalid.
+    # Raised by Methods#api_operation! when the request parameters are invalid.
     class ParametersInvalid < StandardError
+
+      # The parameters.
       attr_reader :params
 
       def initialize(params)
@@ -11,13 +13,14 @@ module Jsapi
         super('')
       end
 
-      # Overrides <code>StandardError#message</code> to lazily generate the error message.
+      # Returns the errors encountered.
+      def errors
+        @params.errors.errors
+      end
+
+      # Overrides <code>Exception#message</code> to lazily generate the error message.
       def message
-        "#{
-          @params.errors.full_messages.map do |message|
-            message.delete_suffix('.')
-          end.join('. ')
-        }."
+        "#{@params.errors.full_messages.map { |m| m.delete_suffix('.') }.join('. ')}."
       end
     end
   end